Корректировка под Code Style
